requirements.txt
pywinauto
comtypes
Pillow
PyYAML
pytest
pytest-html
python-docx
